+++ lib/govuk_publishing_components/engine.rb
@@ -1,3 +1,6 @@
+require "rails/engine"
+require "action_dispatch"
+
 module GovukPublishingComponents
   class Engine < ::Rails::Engine
     isolate_namespace GovukPublishingComponents